requirements.txt
altair==5.4.0
asn1crypto==1.5.1
attrs==24.2.0
awswrangler==3.9.0
blinker==1.8.2
boto3==1.34.160
botocore==1.34.160
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langdetect==1.0.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.4.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pg8000==1.31.2
pillow==10.4.0
protobuf==5.27.3
psycopg2==2.9.9
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
s3transfer==0.10.2
scramp==1.4.5
setuptools==74.1.2
six==1.16.0
smmap==5.0.1
streamlit==1.37.1
streamlit-chat==0.1.1
streamlit-date-picker==0.0.3
tenacity==8.5.0
tiktoken==0.7.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.2